requirements.txt
click==7.1.2
DateTime==4.3
Flask==1.1.2
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.5.1
greenlet==1.0.0
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
PyJWT==2.0.1
pytz==2021.1
six==1.15.0
SQLAlchemy==1.4.7
uuid==1.30
Werkzeug==1.0.1
zope.interface==5.3.0
gunicorn